JUS-123

+++ server/controllers/eventController.js
@@ -21,6 +21,9 @@ const getEventImage = async (req, res, next) => {
 
 const addEvent = async (req, res, next) => {
     const errors = validationResult(req);
+    if (!req.file) {
+        return clientError(res, 'Invalid file provided.');
+    }
     if (!errors.isEmpty()) {
         eventImageService.deleteImage(req.file.path);
         return clientError(res, errors.array());
@@ -102,14 +105,14 @@ const getAnEvent = async (req, res, next) => {
 
 const getOrganizerEvents =  async (req, res, next) => {
     try{
-        Event.find({ 'organizer_id': req.query.id })
+        Event.find({ 'organizerId': req.query.id })
             .exec()
             .then(output => {
                 const response = {
                     count: output.length,
                     events: output.map(out => {
                         return {
-                            id: output._id,
+                            id: out._id, 
                             name: out.eventInfo.name,
                             description: out.eventInfo.description,
                             address:{
@@ -118,7 +121,7 @@ const getOrganizerEvents =  async (req, res, next) => {
                                 country: out.eventInfo.address.country,
                                 postalCode: out.eventInfo.address.postalCode
                             },
-                            eventImagePath: out.eventImage_path,
+                            eventImagePath: out.eventImagePath,
                             bidHistory: out.bidHistory
                         };
                     })
@@ -143,8 +146,7 @@ const updateEvents = async (req, res, next) => {
         else if (event.eventInfo.status == "Ongoing" || event.eventInfo.status == "Completed"){
           return clientError(res, "event cannot be updated");
         }
-        let deletepath = event.eventImage_path
-        console.log(deletepath)
+        let deletepath = event.eventImagePath
         if (req.file){
             const update = {
                 eventInfo: {
